requirements.txt
# Pytorch + Pytorch-Lightning
# Adapt the cuda versions for torch and torchvision to your needs
--find-links https://download.pytorch.org/whl/cu116/torch_stable.html
torch>=1.13.1+cu116
torchvision>=0.14.1+cu116
torchmetrics==1.0.3
pycocotools==2.0.7
lightning==2.0.9
tensorboard>=2.12.0

# Madgrad optimizer
madgrad==1.2

# Hydra
hydra-core==1.3.2
hydra-colorlog==1.2.0
omegaconf==2.3.0

# Packages for data handling, augmentation and visualization
numpy>=1.23.2
opencv-python>=4.6.0.66
albumentations>=1.3.0
pillow>=9.2.0
matplotlib>=3.5.3

# Code Formatter
black>=22.12.0

# Needed for loading .mat files for processing VOC2010_Context Dataset
scipy>=1.9.0

